fix unexpected running=false on mac
Problem is that psprocess is trying to detect pid reuse by looking at
start ime, but is computing startime based on elapased time, which is
unstable.

turns out, this check was unnecessary anyway, because the
psprocess.NewProcess is already checking if the pid is valid & we're not
worried about pid reuse.