Drop Initialized and Disposed timestamps from components.
They are not being used & they are an inhibitor to some refactoring I'm
about to do. We can re-add them if we need them after the new state
management stuff lands.

This change will leave them in the exo state json file, but they are
harmless and will be silently ignored, then dropped if any changes
occur.