eliminate pages directory
Why?

All components in Svelte (and React, and most other frameworks) live in
a single giant namespace. That is, you don't see the module path in the
devtools or in the <TagNames> or anything like that generally. Having it
be a separate directory doesn't really help much at all and just added
to my confusion trying to find components.